+++ src/components/ui/loader/CoveringLoader.stories.tsx
@@ -8,24 +8,51 @@ const meta: Meta = {
     title: "Components/UI/Loader/CoveringLoader",
     component: CoveringLoader,
     tags: ["autodocs", "ui"],
+    args: {
+        show: false,
+        mode: "before",
+    },
+    argTypes: {
+        show: {
+            control: "boolean",
+        },
+        size: {
+            control: {
+                type: "text",
+            },
+        },
+        mode: {
+            options: ["before", "after", "over"],
+            control: { type: "select" },
+        },
+        inertMode: {
+            options: ["block-interaction", "cover-element", "pass-through"],
+            control: { type: "select" },
+        },
+    },
 };
 
 type Story = StoryObj<typeof CoveringLoader>;
 
 const Primary: Story = {
     args: {},
-    render: () => {
+    render: (args) => {
+        const ref = (value: unknown) => { console.info("ref of CoveringLoader:", value); };
+        const innerRef = (value: unknown) => { console.info("children div of CoveringLoader:", value); };
+
         return (
-            <CoveringLoader show={true}>
-                <div style={{ background: "#f0f0f0" }}>
-                    some contents<br />
-                    some contents Lorem ipsum etc
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quibusdam.
-                    some contents<br />
-                    some contents<br />
-                    some contents<br />
-                </div>
-            </CoveringLoader>
+            <div className={"parent"} style={{ background: "red" }}>
+                <CoveringLoader {...args} ref={ref}>
+                    <div style={{ background: "#f0f0f0" }} ref={innerRef}>
+                        some contents<br />
+                        some contents Lorem ipsum etc
+                        Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quibusdam.
+                        some contents<br />
+                        some contents<br />
+                        some contents<br />
+                    </div>
+                </CoveringLoader>
+            </div>
         );
     },
 };